# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /audit:
    post:
      operationId: createAudit
      tags: [audit]
      summary: Submit URL for accessibility audit
      description: Runs an automated accessibility audit on the provided URL and returns a compliance snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuditBody"
      responses:
        "200":
          description: Audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditResult"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Audit failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /audit/{auditId}:
    get:
      operationId: getAudit
      tags: [audit]
      summary: Get audit result by ID
      description: Retrieve a previously run audit result
      parameters:
        - name: auditId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditResult"
        "404":
          description: Audit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /audit/{auditId}/pdf:
    get:
      operationId: getAuditPdf
      tags: [audit]
      summary: Download audit result as PDF
      description: Generates and returns a branded PDF report for a previously run audit
      parameters:
        - name: auditId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF report file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid audit ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Audit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: PDF generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ScanViewportUsed:
      type: object
      properties:
        width:
          type: integer
        height:
          type: integer
        label:
          type: string
      required:
        - width
        - height
        - label
    RuntimeDiagnostics:
      type: object
      properties:
        consoleErrors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              text:
                type: string
            required:
              - type
              - text
        failedRequests:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              errorText:
                type: string
            required:
              - url
      required:
        - consoleErrors
    ScanMetadata:
      type: object
      properties:
        profile:
          type: string
          enum: [default, strict]
        multiViewport:
          type: boolean
        viewportsUsed:
          type: array
          items:
            $ref: "#/components/schemas/ScanViewportUsed"
        runtimeDiagnostics:
          $ref: "#/components/schemas/RuntimeDiagnostics"
      required:
        - profile
        - multiViewport
        - viewportsUsed
    CreateAuditBody:
      type: object
      properties:
        url:
          type: string
          description: The URL to audit for accessibility compliance
        profile:
          type: string
          enum: [default, strict]
          description: |
            `strict` includes additional AAA-oriented axe tags. Default `default` (WCAG 2.x AA-oriented set).
        multiViewport:
          type: boolean
          description: |
            When true, runs axe at mobile and desktop breakpoints and merges results (slower). Default false.
      required:
        - url
    AuditViolationInstance:
      type: object
      description: One DOM node axe flagged for this rule (up to 3 per violation on new audits)
      properties:
        selector:
          type: string
          description: CSS selector chain from axe (joined target segments)
        htmlSnippet:
          type: string
          description: Truncated outer HTML of the element in the scanned page
        failureSummary:
          type: string
          description: Axe human-readable reason this node failed the check
        elementScreenshot:
          type: string
          description: |
            JPEG data URL of the node's rendered bounds (Playwright engine only; omitted when capture fails).
        checkDetails:
          type: array
          items:
            type: string
          description: Messages from axe check results attached to this node (`any` / `all` / `none`).
      required:
        - selector
        - htmlSnippet
    AuditViolation:
      type: object
      properties:
        id:
          type: string
          description: WCAG criterion ID (e.g. "color-contrast")
        wcagCriteria:
          type: string
          description: WCAG criterion reference (e.g. "1.4.3 Contrast Minimum")
        description:
          type: string
        impact:
          type: string
          enum: [minor, moderate, serious, critical]
        affectedElements:
          type: integer
          description: Number of affected elements on the page
        topSelectors:
          type: array
          items:
            type: string
          description: Up to 3 representative CSS selectors for affected elements
        help:
          type: string
          description: Short remediation guidance from axe-core for this rule.
        helpUrl:
          type: string
          description: Link to axe rule documentation (Deque University).
        instanceDetails:
          type: array
          items:
            $ref: "#/components/schemas/AuditViolationInstance"
          description: |
            Up to 3 affected nodes with selector, HTML, failure text, optional element screenshot, and axe check messages.
            Omitted on legacy audits. Remote scans do not include original source file line numbers; use DevTools with the selector.
        detectedInViewports:
          type: array
          items:
            type: string
          description: Breakpoint labels when this finding was merged from a multi-viewport scan.
      required:
        - id
        - wcagCriteria
        - description
        - impact
        - affectedElements
        - topSelectors
    AuditResult:
      type: object
      properties:
        auditId:
          type: string
        url:
          type: string
        scannedAt:
          type: string
          format: date-time
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall accessibility compliance score
        level:
          type: string
          enum: [critical, poor, moderate, good, excellent]
          description: Score level label
        totalViolations:
          type: integer
        criticalViolations:
          type: integer
        seriousViolations:
          type: integer
        violations:
          type: array
          items:
            $ref: "#/components/schemas/AuditViolation"
        passedChecks:
          type: integer
        totalChecks:
          type: integer
        scanEngine:
          type: string
          enum: [playwright, static_fallback, unknown]
          description: |
            Engine used for this audit: `playwright` is headless Chromium with axe (full JS/DOM).
            `static_fallback` is HTML fetched into JSDOM when the browser engine fails (no client-side JS).
            `unknown` is stored for legacy rows before this field existed.
        pageScreenshot:
          type: string
          description: |
            Viewport JPEG data URL of the page after the axe run (Playwright only). Omitted for static fallback, legacy rows, or when capture failed.
        scanMetadata:
          $ref: "#/components/schemas/ScanMetadata"
          description: Scan options, viewports used, and optional runtime diagnostics (Playwright). Omitted on legacy rows.
      required:
        - auditId
        - url
        - scannedAt
        - score
        - level
        - totalViolations
        - criticalViolations
        - seriousViolations
        - violations
        - passedChecks
        - totalChecks
        - scanEngine
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message